+++ src/schedule/__tests__/algorithm.test.ts
@@ -1,8 +1,7 @@
-import { generateSchedule } from "../../../schedule/utils"
-import { oneEventCal, threeEventCal, twoEventCal } from "../../../examples/calendarExamples";
-import { preferenceExample } from "../../../examples/preferenceExample";
-import { oneJob } from "../../../examples/oneJob";
-
+import { generateSchedule } from "../utils";
+import { oneEventCal, threeEventCal, twoEventCal } from "./calendarExamples";
+import { oneJob } from "./oneJob";
+import { preferenceExample } from "./preferenceExample";
 
 it("Test scheduler one event and one job", async () => {
     const jobs = [oneJob];
@@ -13,8 +12,8 @@ it("Test scheduler one event and one job", async () => {
 
     const events = schedule.jCal[2].slice(1);
 
-    expect(String(events[0][1][1][3])).toMatch("2021-05-11T13:00:00");
-    expect(String(events[0][1][2][3])).toMatch("2021-05-11T14:00:00");
+    expect(String(events[0][1][1][3])).toMatch("2021-05-11T01:00:00");
+    expect(String(events[0][1][2][3])).toMatch("2021-05-11T02:00:00");
 });
 
 it("Test scheduler two events and one job", async () => {
@@ -25,9 +24,9 @@ it("Test scheduler two events and one job", async () => {
     let schedule: any = generateSchedule(calendar, jobs, preferences);
 
     const events = schedule.jCal[2].slice(1);
-    
-    expect(String(events[0][1][1][3])).toMatch("2021-05-11T15:00:00");
-    expect(String(events[0][1][2][3])).toMatch("2021-05-11T16:00:00");
+
+    expect(String(events[0][1][1][3])).toMatch("2021-05-11T04:00:00");
+    expect(String(events[0][1][2][3])).toMatch("2021-05-11T05:00:00");
 });
 
 it("Test scheduling job between events", async () => {
@@ -38,7 +37,7 @@ it("Test scheduling job between events", async () => {
     let schedule: any = generateSchedule(calendar, jobs, preferences);
 
     const events = schedule.jCal[2].slice(1);
-    
-    expect(String(events[0][1][1][3])).toMatch("2021-05-11T15:00:00");
-    expect(String(events[0][1][2][3])).toMatch("2021-05-11T16:00:00");
+
+    expect(String(events[0][1][1][3])).toMatch("2021-05-11T03:00:00");
+    expect(String(events[0][1][2][3])).toMatch("2021-05-11T04:00:00");
 });
