+++ src/schedule/__tests__/calendarExamples.ts
@@ -58,8 +58,8 @@ export const oneEventCal = {
                 "vevent",
                 [
                     ["summary", {}, "text", "Event 1"],
-                    ["dtstart", {}, "date", "2021-05-11T00:00:00.000Z"],
-                    ["dtend", {}, "date", "2021-05-11T01:00:00.000Z"],
+                    ["dtstart", {}, "date", "2021-05-11T00:00:00.000"],
+                    ["dtend", {}, "date", "2021-05-11T01:00:00.000"],
                 ],
                 [],
             ],
@@ -127,8 +127,8 @@ export const twoEventCal = {
                 "vevent",
                 [
                     ["summary", {}, "text", "Event 1"],
-                    ["dtstart", {}, "date", "2021-05-11T00:00:00.000Z"],
-                    ["dtend", {}, "date", "2021-05-11T03:00:00.000Z"],
+                    ["dtstart", {}, "date", "2021-05-11T00:00:00.000"],
+                    ["dtend", {}, "date", "2021-05-11T03:00:00.000"],
                 ],
                 [],
             ],
@@ -136,8 +136,8 @@ export const twoEventCal = {
                 "vevent",
                 [
                     ["summary", {}, "text", "Event 2"],
-                    ["dtstart", {}, "date", "2021-05-04T01:00:00.000Z"],
-                    ["dtend", {}, "date", "2021-05-04T03:00:00.000Z"],
+                    ["dtstart", {}, "date", "2021-05-11T03:00:00.000"],
+                    ["dtend", {}, "date", "2021-05-11T04:00:00.000"],
                 ],
                 [],
             ],
@@ -205,8 +205,8 @@ export const threeEventCal = {
                 "vevent",
                 [
                     ["summary", {}, "text", "Event 1"],
-                    ["dtstart", {}, "date", "2021-05-11T00:00:00.000Z"],
-                    ["dtend", {}, "date", "2021-05-11T01:00:00.000Z"],
+                    ["dtstart", {}, "date", "2021-05-11T00:00:00.000"],
+                    ["dtend", {}, "date", "2021-05-11T01:00:00.000"],
                 ],
                 [],
             ],
@@ -214,8 +214,8 @@ export const threeEventCal = {
                 "vevent",
                 [
                     ["summary", {}, "text", "Event 2"],
-                    ["dtstart", {}, "date", "2021-05-11T01:00:00.000Z"],
-                    ["dtend", {}, "date", "2021-05-11T03:00:00.000Z"],
+                    ["dtstart", {}, "date", "2021-05-11T01:00:00.000"],
+                    ["dtend", {}, "date", "2021-05-11T03:00:00.000"],
                 ],
                 [],
             ],
@@ -223,8 +223,8 @@ export const threeEventCal = {
                 "vevent",
                 [
                     ["summary", {}, "text", "Event 3"],
-                    ["dtstart", {}, "date", "2021-05-11T04:00:00.000Z"],
-                    ["dtend", {}, "date", "2021-05-11T05:00:00.000Z"],
+                    ["dtstart", {}, "date", "2021-05-11T04:00:00.000"],
+                    ["dtend", {}, "date", "2021-05-11T05:00:00.000"],
                 ],
                 [],
             ],
